Use the inferiors auxiliary vector to resolve addresses for PIE's.
Position Independent Executables are loaded into virtual memory at a random
offset.  This offset can not be statically known and needs to be computed using
information in the inferiors loaded image.

On Linux, processes are brought up with an auxiliary vector on the stack.
Among other things, this vector contains the (dynamic) entry address of the
process.  One does not need to walk the stack to find this information as it is
also available under /proc/<pid>/auxv.

The new AuxVector class provides a convenient read-only view of the auxiliary
vector information.  We use the dynamic entry address and the address as
specified in the object file to compute the actual load address of the
inferior.  This strategy works for both normal executables and PIE's.