Patches from Jean-Daniel:
One (stepout.patch) to fix a problem in ThreadPlanStepOut.cpp. There is an erroneous semi colon at end of an if statement that make the condition useless (if body is empty).

And the second patch is to remove to useless typedef on enum, and so avoid a lot of warnings with clang++.


git-svn-id: http://llvm.org/svn/llvm-project/lldb/trunk@124874 91177308-0d34-0410-b5e6-96231b3b80d8